refactor(Pawn, PawnTest): 파일 구조 변경
기존 domain 폴더 내에 있던 Pawn.java를 이동
