+++ src/main/java/softeer2nd/Pawn.java
@@ -1,4 +1,4 @@
-package softeer2nd.domain;
+package softeer2nd;
 
 public class Pawn {
