+++ src/test/java/softeer2nd/PawnTest.java
@@ -1,4 +1,4 @@
-package softeer2nd.domain;
+package softeer2nd;
 
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
